==> 细胞焦亡联合单细胞文章复现/4_tcga_model/1_surv_dat.Rmd <==
---
title: "生存分析前的数据整理"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
knitr::opts_chunk$set(fig.width = 6,fig.height = 6,collapse = TRUE)
knitr::opts_chunk$set(message = FALSE)
```

与上课代码的主要区别：

1.使用了TPM而非CPM
2.有TNM信息的拆分


表达矩阵只需要tumor数据，不要normal，将其去掉，新表达矩阵数据命名为exprSet；

临床信息需要进一步整理，成为生存分析需要的格式，新临床信息数据命名为meta。

由于不同癌症的临床信息表格列名可能不同，这里的代码需要根据实际情况修改。

```{r}
rm(list=ls())
proj = "TCGA-STAD"
load(paste0("../2_tcga_stad/",proj,".Rdata"))
library(stringr)
```

### 1.整理表达矩阵

#### 1.1去除normal样本

```{r}
table(Group)
exprSet = exp[,Group=='tumor']
ncol(exp)
ncol(exprSet)
```


#### 1.2 基因过滤

再次进行基因过滤.

(1)标准1：至少要在50%的样本里表达量大于0（最低标准）。

exp(600样本)满足“至少在300个样本里表达量>0”,不能等同于
exprSet(500样本)满足“至少在250个样本里表达量>0”

```{r}
k = apply(exprSet,1, function(x){sum(x>0)>0.5*ncol(exprSet)});table(k)
exprSet = exprSet[k,]
nrow(exprSet)
```

(2)标准2：至少在一半以上样本里表达量＞10(其他数字也可，酌情调整)

```{r}
k = apply(exprSet,1, function(x){sum(x>10)>0.5*ncol(exprSet)});table(k)
exprSet = exprSet[k,]
nrow(exprSet)
```

#### 1.3 使用logCPM或logTPM数据
这里是用了tmp。

```{r}
#cpm
#exprSet = log2(edgeR::cpm(exprSet)+1)
#tmp
exprSet[1:4,1:4]
load("stad_tpm.Rdata")
stad_tpm[1:4,1:4]
library(tinyarray)
stad_tpm = trans_exp_new(stad_tpm)
stad_tpm = log2(stad_tpm+1)
stad_tpm = stad_tpm[rownames(exprSet),colnames(exprSet)]
exprSet= stad_tpm
exprSet[1:4,1:4]
boxplot(exprSet[,1:10])
```

#### 1.4 样本与病人

有的病人会有两个或两个以上的肿瘤样本，就有重复。两种可行的办法：

（1）以病人为中心，对表达矩阵的列按照病人ID去重复，每个病人只保留一个样本。

```{r}
exprSet = exprSet[,sort(colnames(exprSet))]
k = !duplicated(str_sub(colnames(exprSet),1,12));table(k)
exprSet = exprSet[,k] 
ncol(exprSet)
```

（2）以样本为中心，如果每个病人有多个样本则全部保留。(删掉上面这一段代码即可)

### 2.整理生存信息和临床信息

#### 2.1数据合并

xena单独整理了生存信息，把它和临床信息合并到一起。

```{r}
library(dplyr)
surv = read.delim("TCGA-STAD.survival.tsv")
meta = left_join(surv,clinical,by = c("X_PATIENT"= "bcr_patient_barcode"))
nrow(meta)
length(unique(meta$sample))
meta = distinct(meta,sample,.keep_all = T)
```

#### 2.2 样本过滤

去掉生存信息不全或者生存时间小于30天的样本，样本纳排标准不唯一，且差别很大

```{r}
k1 = meta$OS.time >= 30;table(k1)
k2 = !(is.na(meta$OS.time)|is.na(meta$OS));table(k2)
meta = meta[k1&k2,]
```

#### 2.3 选列、简化列名

小技巧：搜索列名

```{r}
tmp = data.frame(colnames(meta))
#View(tmp)
```

选择需要的列，简化列名

```{r}
meta = meta[,c(
  'sample',
  'OS',
  'OS.time',
  'race_list',
  'age_at_initial_pathologic_diagnosis',
  'gender' ,
  'stage_event_pathologic_stage',
  'stage_event_tnm_categories'
)]
colnames(meta)=c('ID','event','time','race','age','gender','stage','TNM')
str(meta)
meta$gender = as.character(meta$gender)
meta$stage = as.character(meta$stage)
meta$race = as.character(meta$race)
meta$TNM = as.character(meta$TNM)
```

#### 2.4 简化、规范内容

(1) 结局事件

生存分析的输入数据里，要求结局事件必须用0和1表示，1表示阳性结局。

xena的数据是整理好的，其他来源的需要自行检查和整理。

```{r}
table(meta$event)
#自行整理的例子：
a = c("Dead", "Dead", "Alive", "Dead", "Alive")
ifelse(a=="Alive",0,1)
```

(2) 生存时间

认清生存时间的单位（通常是月，也可以用天和年）;

```{r}
range(meta$time)
meta$time = meta$time/30
range(meta$time)
```

(3)检查各列的信息是否规范

没有冗余信息

缺失的信息用NA代替

```{r}
meta[meta==""] = NA
table(meta$stage)
meta$stage = meta$stage %>% 
  str_remove_all("Stage |A|B|C") 
table(meta$stage,useNA = "always")
table(meta$race,useNA = "always")
table(meta$gender,useNA = "always")
```

# tnm拆分

```{r}
#特殊情况：没有N的，手动加上NX
table(meta$TNM %>% str_detect("N"))
n = which(! meta$TNM %>% str_detect("N"))
meta$TNM[n]
meta$TNM[n] =c("T3NXM0", "T3NXM0", "T4NXM0")

tnm = meta$TNM %>% 
  str_remove_all("a|b|c") %>%
  str_split("T|N|M",simplify = T) %>% 
  as.data.frame() %>% select(-1) 
head(tnm)
colnames(tnm) = c("T","N","M")
table(tnm$M,useNA = "always")
tnm = apply(tnm,2,as.numeric) # 转换数值型，X无法转换，正好变成NA
meta = cbind(meta[,-8],tnm)
```

```{r}
str(meta)
```

### 3.实现表达矩阵与临床信息的匹配 

即：meta的每一行与exprSet每一列一一对应

```{r}
rownames(meta) <- meta$ID
head(rownames(meta))
head(colnames(exprSet))
colnames(exprSet) = str_sub(colnames(exprSet),1,16)
s = intersect(rownames(meta),colnames(exprSet));length(s)
exprSet = exprSet[,s]
meta = meta[s,]
dim(exprSet)
dim(meta)
identical(rownames(meta),colnames(exprSet))
save(meta,exprSet,proj,file = paste0(proj,"_sur_model.Rdata"))
```
