==> 细胞焦亡联合单细胞文章复现/4_tcga_model/3_lasso.Rmd <==
---
title: "lasso回归-生存模型专用"
author: "生信技能树"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
knitr::opts_chunk$set(fig.width = 7, fig.height = 5,collapse = TRUE)
knitr::opts_chunk$set(message = FALSE)
```

选择构建模型的基因：是marker基因和deg的交集

### 1.准备输入数据

```{r}
rm(list = ls())
proj = "TCGA-STAD"
load(paste0(proj,"_sur_model.Rdata"))
load(paste0(proj,"_logrank_cox_gene.Rdata"))
load("../2_tcga_stad/gs.Rdata") #差异基因和单细胞marker基因的交集，在../2_tcga_stad/3.enrich.Rmd里面生成的
load("../3_GEO_dat_pre/geoexp.Rdata")  #GEO表达矩阵
library(tinyarray)
g = intersect_all(gs,cox,rownames(exp)) ;length(g)
exprSet = exprSet[g,]
```

### 2.构建lasso回归模型

输入数据是表达矩阵(仅含tumor样本)和每个病人对应的生存资料（顺序必须一致）。

```{r}
library(glmnet)
library(survival)
x = t(exprSet) #转置后的基因表达矩阵
y = Surv(meta$time,meta$event) #生存信息
set.seed(110)
cvfit = cv.glmnet(x, y, family="cox") #10折交叉验证，用于调优参数
plot(cvfit)
fit = glmnet(x, y, family = "cox") #建模
plot(fit,xvar="lambda",label = F)
coef = coef(fit,s = cvfit$lambda.min) #提取最优lamda参数对应的模型的基因系数
class(coef) #稀疏矩阵
dat = data.frame(gene = rownames(coef),
                 coefficient = as.numeric(coef[,1]))
head(dat)
dat = dat[dat$coefficient!=0,]
nrow(dat)
lassoGene = dat$gene
save(lassoGene,file = paste0(proj,"_lassogene.Rdata"))
```

### 3.模型预测和评估

> C-index用于计算生存分析中的COX模型预测值与真实之间的区分度（discrimination），也称为Harrell’s concordanceindex。C-index在0.5-1之间。0.5为完全不一致,说明该模型没有预测作用,1为完全一致,说明该模型预测结果与实际完全一致。

```{r}
lasso.prob <- predict(fit, newx = x ,s = cvfit$lambda.min)
library(Hmisc)
options(scipen=200)
rcorr.cens(-lasso.prob[,1],
           Surv(meta$time, meta$event))
```
