==> 细胞焦亡联合单细胞文章复现/4_tcga_model/4_cox-forest.Rmd <==
---
title: "cox-forest"
author: "生信技能树"
output: rmarkdown::html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
knitr::opts_chunk$set(fig.width = 7,fig.height = 5, collapse = TRUE,message = F,warning = F)
```

### 1.准备输入数据

```{r}
rm(list = ls())
proj = "TCGA-STAD"
load(paste0(proj,"_sur_model.Rdata"))
load(paste0(proj,"_lassogene.Rdata"))
g = lassoGene
exprSet = exprSet[g,]
```

### 2.构建coxph模型 

将用于建模的基因（例如lasso回归选中的基因）从表达矩阵中取出来，，可作为列添加在meta表噶的后面,组成的数据框赋值给dat。

```{r}
library(stringr)
e=t(exprSet)
colnames(e)= str_replace_all(colnames(e),"-","_")
dat=cbind(meta,e) 

dat$gender=as.numeric(factor(dat$gender))
dat$stage=as.numeric(factor(dat$stage))
colnames(dat)
```

```{r}
library(survival)
library(survminer)
model = coxph(formula = Surv(time, event) ~ ., data = dat[, -c(1, 4:10)])
#波浪号后面的点“.”代表除了time和event外的其他所有列
#这一步仅使用基因，不需要临床信息，所以去掉了ID和临床信息列
```

逐步回归法可以帮你拿到最优模型

```{r,results='hide'}
model2 = step(coxph(formula = Surv(time, event) ~ ., data = dat[, -c(1, 4:10)]))
#如果决定使用此结果就删掉#👇
#model = model2
```


#### 风险森林图

```{r}
ggforest(model,data = dat)
```

### 4.模型预测和评估

```{r }
#计算预测值，基因表达量的加权求和，apply实现了批量计算，所有样本都计算。
fp = apply(dat[,names(model$coefficients)], 1, function(x){
  sum(x * model$coefficients)
})
head(fp)
library(Hmisc)
options(scipen=200)
rcorr.cens(-fp,Surv(dat$time, dat$event))
#将风险值和风险分组作为一列加到dat里面
dat$Riskscore = fp
dat$Risk = ifelse(fp<median(fp),"low_risk","high_risk")
dat$Risk = factor(dat$Risk,levels = c("low_risk","high_risk"))
save(dat,model,fp,file = paste0(proj,"cox_fp.Rdata"))
```

#### 划分高低风险并画KM-plot

```{r}
sfit <- survfit(Surv(time, event)~Risk, data=dat)

ggsurvplot(sfit,palette = "jco",
           risk.table =TRUE,pval =TRUE,xlab ="Time in months", 
           ggtheme =theme_classic())
```

#### 风险因子三图联动

```{r}
library(tinyarray)
#tinyarray里现成的函数
risk_plot(exprSet_hub = t(dat[,names(model$coefficients)]),
          meta = meta,
          riskscore = fp)
packageVersion("tinyarray")
#在2.4.1以上的新版本里添加了一个参数n_cutoff,默认值是3，意思是把热图颜色范围设置为-3~3之间，颜色会更鲜明，代码不变，更新一下颜值更高😄
```
