==> 细胞焦亡联合单细胞文章复现/4_tcga_model/5_risk_plot.Rmd <==
---
title: "三图联动"
author: "生信技能树"
output: rmarkdown::html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
knitr::opts_chunk$set(fig.width = 8, collapse = TRUE)
knitr::opts_chunk$set(message = FALSE,fig.height = 10,warning = FALSE)
```
### 1.准备输入数据

```{r}
rm(list = ls())
library(ggplot2)
library(ggplotify)
proj = "TCGA-STAD"
load(paste0(proj,"_sur_model.Rdata"))
load(paste0(proj,"cox_fp.Rdata"))
```

### 2.风险因子三图联动

```{r}
fp_dat=data.frame(patientid=1:length(fp),
                  fp=as.numeric(sort(fp)),
                  ri= dat$Risk[order(fp)])

sur_dat=data.frame(patientid=1:length(fp),
                   time=dat[order(fp),'time'] ,
                   event=dat[order(fp),'event']) 

sur_dat$event=ifelse(sur_dat$event==0,'alive','death')
sur_dat$event=factor(sur_dat$event,levels = c("death","alive"))

library(stringr)
gs = names(model$coefficients)[names(model$coefficients)%in% rownames(exprSet)]
gs
exp_dat = t(dat[order(fp),gs])
colnames(exp_dat) = str_sub(colnames(exp_dat),1,12)
rownames(exp_dat) = str_replace_all(rownames(exp_dat),"_","-")
###第一个图----
p1=ggplot(fp_dat,aes(x=patientid,y=fp))+
  geom_point(aes(color = ri))+
  scale_color_manual(values = c("#2874C5", "#f87669"))+
  geom_vline(xintercept = 0.5*nrow(sur_dat),lty = 2)+
  scale_x_continuous(expand=c(0.01,0))+
  labs(x = "",y = "risk score")+
  theme_bw()
#第二个图
p2=ggplot(sur_dat,aes(x=patientid,y=time))+
  geom_point(aes(col=event))+
  scale_color_manual(values = c("#f87669","#2874C5"))+
  geom_vline(xintercept = 0.5*nrow(sur_dat),lty = 2)+
  scale_x_continuous(expand=c(0.01,0))+
  labs(x = "")+
  theme_bw()
#第三个图
annotation_col = dat["Risk"]
mycolors <- colorRampPalette(c("#2874C5","white", "#f87669"), bias = 1.2)(100)
ann_colors = list(
    group = c(lowrisk="#2874C5", highrisk="#f87669")
)
p3=pheatmap::pheatmap(exp_dat,
                      col= mycolors,
                      annotation_col = annotation_col,
                      annotation_colors =ann_colors,
                      scale = "row",
                      breaks = seq(-3,3,length.out = 100),
                      show_colnames = F,
                      cluster_cols = F,
                      annotation_legend = F)
#拼图实现三图联动
library(patchwork)
p1/p2/as.ggplot(p3)

library(tinyarray)
n = scale(t(exp_dat))
n[n > 3] = 3
n[n < -3] = -3
p3 = ggheat(n,fp_dat$ri,show_rownames = F)+
  theme(axis.text = element_text(size = 8))
p3
p1 /p2 /p3 + plot_layout(design = 'A
                         B
                         C
                         C
                         C
                         C')
ggsave("risk_plot.png",height = 10,width = 10)
```

从上向下三个图分别表示:

1.每个病人的预测值，按照从小到大排序
2.每个病人的生存时间，颜色区分生死
3.热图，挑出的基因在每个样本中的表达量
