==> 细胞焦亡联合单细胞文章复现/4_tcga_model/6_timeROC.Rmd <==
---
title: "timeROC"
author: "生信技能树"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
knitr::opts_chunk$set(fig.width = 7, fig.height = 5,collapse = TRUE)
knitr::opts_chunk$set(message = FALSE)
```

### 1.构建模型并完成预测

```{r}
rm(list = ls())
proj = "TCGA-STAD"
load(paste0(proj,"cox_fp.Rdata"))
```

#### 2 time-ROC计算和出图

```{r}
library(survminer)
library(survival)
library(timeROC)
result <-with(dat, timeROC(T=time,
                     delta=event,
                     marker=Riskscore,
                     cause=1,
                     times=c(12,36,60),
                     iid = TRUE))
plot_dat = data.frame(fpr = as.numeric(result$FP),
                 tpr = as.numeric(result$TP),
                 time = rep(as.factor(c(12,36,60)),each = nrow(result$TP)))

library(ggplot2)
ggplot() + 
  geom_line(data = plot_dat,aes(x = fpr, y = tpr,color = time),linewidth = 1) + 
  scale_color_manual(name = NULL,values = c("#2874C5", "#f87669", "#e6b707"),
                     labels = paste0("AUC of ",c(1,3,5),"-y survival: ",
                                     format(round(result$AUC,2),nsmall = 2)))+
  geom_line(aes(x=c(0,1),y=c(0,1)),color = "grey")+
  theme_bw()+
  theme(panel.grid = element_blank(),
        legend.background = element_rect(linetype = 1, linewidth = 0.2, colour = "black"),
        legend.position = c(0.765,0.125))+
  scale_x_continuous(expand = c(0.005,0.005))+
  scale_y_continuous(expand = c(0.005,0.005))+
  labs(x = "1 - Specificity",
       y = "Sensitivity")+
  coord_fixed()
```
