==> 细胞焦亡联合单细胞文章复现/4_tcga_model/7.nomo.Rmd <==
---
title: "模型可视化"
author: "生信技能树"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,message = F,warning = F)
```

### 1.森林图

```{r}
rm(list = ls())
proj = "TCGA-STAD"
load(paste0(proj,"_sur_model.Rdata"))
load(paste0(proj,"cox_fp.Rdata"))

library(stringr)
library(survival)
library(survminer)
```

### 2.诺模图

```{r fig.width= 10,fig.height= 7}
library(rms)
dat$time = dat$time*30
#colnames(dat)[6:7]
dd<-datadist(dat)
options(datadist="dd")
#套用代码时需要修改下面这行公式~后面的部分
mod <- cph(Surv(time, event) ~ age + gender + stage + Risk,
           data=dat,x=T,y=T,surv = T)

surv<-Survival(mod) 
m1<-function(x) surv(365,x)
m3<-function(x) surv(1095,x)
m5<-function(x) surv(1825,x)

x<-nomogram(mod,
            fun = list(m1,m3,m5),
            funlabel = c('1-y survival',
                         '3-y survival',
                         '5-y survival'),
            lp = F)

plot(x)
```

#### 诺莫图校准曲线

```{r}

f1 <- cph(formula =  Surv(time, event) ~ age + gender + stage + Risk,
          data=dat,x=T,y=T,surv = T, time.inc=365)
cal1<- calibrate(f1, cmethod="KM", method="boot", u=365, m=50, B=1000)

f3 <- cph(formula =  Surv(time, event) ~ age + gender + stage + Risk,
          data=dat,x=T,y=T,surv = T, time.inc=1095)
cal3 <- calibrate(f3, cmethod="KM", method="boot", u=1095, m=50, B=1000)

f5 <- cph(formula =  Surv(time, event) ~ age + gender + stage + Risk,
          data=dat,x=T,y=T,surv = T, time.inc=1825)
cal5 <- calibrate(f5, cmethod="KM", method="boot", u=1825, m=50, B=1000)

data1 <- data.frame(
  Time = rep("1-year", nrow(cal1)),
  Predicted = cal1[,"mean.predicted"],
  Observed = cal1[,"KM"]
)

data3 <- data.frame(
  Time = rep("3-year", nrow(cal3)),
  Predicted = cal3[,"mean.predicted"],
  Observed = cal3[,"KM"]
)

data5 <- data.frame(
  Time = rep("5-year", nrow(cal5)),
  Predicted = cal5[,"mean.predicted"],
  Observed = cal5[,"KM"]
)

all_data <- rbind(data1, data3, data5)
library(ggplot2)
ggplot(all_data, aes(x = Predicted, y = Observed)) +
  geom_abline(intercept = 0, slope = 1, linewidth = 2,linetype = 3, color = "darkgrey") +
  geom_line(aes(color = Time)) + 
  scale_color_manual(values = c("#e6b707", "#2166AC", "#B2182B")) + 
  
  geom_point(aes(shape =Time,fill = Time),
             color = "white",size = 5,stroke = 2) +
  scale_shape_manual(values = c(21,22,23))+
  scale_fill_manual(values = c("#e6b707", "#2166AC", "#B2182B")) + 
  labs(title = "Calibration Curves for Nomogram-Predicted Survival",
       x = "Predicted Probability",
       y = "Observed Probability",
       color = "Time") +
  xlim(0, 1) + ylim(0, 1) +
  theme_classic() + 
  coord_fixed()+
  theme(legend.title = element_text(size = 14),
        legend.text = element_text(size = 12)) 
```
